Feature: allow user to create attributes as non filter attributes.
A new property (is_filter) has been introduced into ProductAttribute,
it's a boolean field generating a checkbox to uncheck (or not) to hide
the given attribute from filtering.

This feature as for goal to make it easier to manage product attributes,
without imposing a filtering or without imposing
to go through the product description which is making things messy.